// src/types/commonType.ts
export interface ITodayExpense {
  icon: string;
  name: string;
  details: string;
  amount: string;
}

export interface IExpense {
  icon: string;
  name: string;
  details: string;
  amount: string;
}

export interface IExpenseCategories {
  name: string;
  amount: string;
  percentage: string;
}